+++ src/app/pages/authentication/login/login.page.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { ToastController } from '@ionic/angular';
 
 // services
 import { AuthService } from '../../../services/auth.service';
@@ -7,30 +8,63 @@ import { EventsService } from '../../../services/events-service';
 
 import { LoginComponent } from '../../../components/authentication/login/login.component';
 
+// utils
+import { isInArray } from 'src/app/utils/utils';
+
 @Component({
   selector: 'app-login',
   templateUrl: './login.page.html',
   styleUrls: ['./login.page.scss'],
 })
-export class LoginPage implements OnInit {
 
+export class LoginPage implements OnInit {
   showSpinnerInLoginBtn = false;
+  showErrorSignIn = false;
   companyLogoBlackUrl: string;
   companyName: string;
+
   public translationMap: Map<string, string>;
+  private subscriptions = [];
+
 
   constructor(
     public authService: AuthService,
     private translateService: CustomTranslateService,
     private events: EventsService,
-    private loginComponent: LoginComponent
+    private loginComponent: LoginComponent,
+    public toastController: ToastController
   ) { }
 
   ngOnInit() {
+    this.initialize();
+  }
+
+
+  /** */
+  ionViewDidEnter() {
+  }
+
+  /** */
+  ionViewWillLeave() {
+    this.unsubescribeAll();
+  }
+
+  /** */
+  initialize() {
     this.companyLogoBlackUrl = 'assets/chat21-logo.svg';
     this.companyName = 'Tiledesk'; // this.chatManager.getTenant();
     this.translations();
     this.events.subscribe('sign-in', this.signIn);
+    this.setSubscriptions();
+  }
+
+  /** */
+  private setSubscriptions() {
+    const keySubscription = 'sign-in';
+    if (!isInArray(keySubscription, this.subscriptions)) {
+      this.subscriptions.push(keySubscription);
+      this.events.subscribe(keySubscription, this.signIn);
+    }
   }
 
   /**
@@ -42,10 +76,11 @@ export class LoginPage implements OnInit {
     console.log('************** signIn', user);
     console.log('************** error', error);
     if (error) {
-      this.showSpinnerInLoginBtn = false;
-      console.log('************** showSpinnerInLoginBtn', this.showSpinnerInLoginBtn);
       // faccio uscire alert
-      this.loginComponent.test();
+      const errore =  this.translationMap.get('LABEL_SIGNIN_ERROR');
+      this.showSpinnerInLoginBtn = false;
+      // this.presentToast(errore);
+      this.loginComponent.showErrorSignIn(errore);
     }
   }
 
@@ -61,10 +96,13 @@ export class LoginPage implements OnInit {
       'LABEL_DONT_HAVE_AN_ACCOUNT_YET',
       'LABEL_SIGNUP',
       'LABEL_FORGOT_YOUR_PASSWORD',
-      'LABEL_CLICK_HERE'
+      'LABEL_CLICK_HERE',
+      'LABEL_SIGNIN_ERROR'
     ];
     this.translationMap = this.translateService.translateLanguage(keys);
   }
+
+
 
   /**
    *
@@ -73,6 +111,28 @@ export class LoginPage implements OnInit {
   returnSignInWithEmailAndPassword(auth: any) {
     console.log('returnSignInWithEmailAndPassword', auth, auth.email, auth.password );
     this.authService.signInWithEmailAndPassword(auth.email, auth.password);
+  }
+
+  /** */
+  // async presentToast(error: string) {
+  //   const toast = await this.toastController.create({
+  //     message: error,
+  //     duration: 2000,
+  //     header: 'Attenzione',
+  //     position: 'top',
+  //     buttons: null
+  //   });
+  //   toast.present();
+  // }
+
+  /** */
+  private unsubescribeAll() {
+    console.log('unsubescribeAll: ', this.subscriptions);
+    this.subscriptions.forEach((subscription: any) => {
+      console.log('unsubescribe: ', subscription);
+      this.events.unsubscribe(subscription, null);
+    });
+    this.subscriptions = [];
   }