+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # chat21-ionic ver 1.0
 
+### 3.0.11
+- new: set persistence firebase from environment 
+- new: get JWT token from url queryParams and signin with token 
+- new: save token in localstorage
+
 ### 3.0.10
 - bug-fix: url navigation
 - bug-fix: CONTACTS_URL from environment